% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{character}

\item{pkgname}{character}
}
\description{
.onAttach
}
\references{
https://r-pkgs.org/r.html
}
